+++ src/rex.deploy/NEWS.rst
@@ -12,6 +12,7 @@
 * Include links to the default selector.
 * Moved ``rex deploy`` and other commands from ``rex.ctl``.
 * Added ``json`` data type.
+* Made ``Model`` an extension.
 
 
 2.0.0 (2015-01-05)